Moved SortableTable component to table folder.
